+++ webapp/package.json
@@ -32,6 +32,7 @@
     "video.js": "^7.15.4"
   },
   "devDependencies": {
+    "@types/js-cookie": "^3.0.2",
     "@types/node": "^14.0.0",
     "@types/react": "^17.0.2",
     "@types/react-dom": "^17.0.2",
